Change variable initialisation to scale better, add deleteset route
Some variables were initialised based on a specific index of a queried
list. This will not scale well once question sets are deleted and added
in as the index of a set in a list will no longer match up with its ID.

The new route, 'deleteset' handles deletion of question sets.